refactor: restructure tft-price pallet #735
